Fix name conflict in Lexers, also fix some warnings.

+++ filesys.cpp
@@ -196,7 +196,7 @@ unsigned char* ReadBinaryFile(const char* fileName, int* outLength) {
 	fseek(fIn, 0, SEEK_SET);
 
 	unsigned char* data = (unsigned char*)malloc(fileLength);
-	int bytesRead = fread(data, 1, fileLength, fIn);
+	int bytesRead = (int)fread(data, 1, fileLength, fIn);
 	ASSERT(bytesRead == fileLength);
 
 	*outLength = fileLength;
@@ -218,7 +218,7 @@ char* ReadTextFile(const char* fileName, int* outLength) {
 	fseek(fIn, 0, SEEK_SET);
 
 	char* data = (char*)malloc(fileLength + 1);
-	int bytesRead = fread(data, 1, fileLength, fIn);
+	int bytesRead = (int)fread(data, 1, fileLength, fIn);
 	ASSERT(bytesRead == fileLength);
 	data[fileLength] = '\0';
